Add Ruby declaration to the Gemfile for heroku
While it's undesriable (in my mind) to peg our
skeleton to a certain Ruby release, heroku gives a
student-alerting error if this declaration is NOT
there.

Given that DBC machines are *not* going to be
running non-standard Rubies, and given that those
who are running non-standard Rubies should know
how to adjust this, I think this is a valid
change.